+++ packages/cli/scripts/bootstrap.sh
@@ -9,8 +9,8 @@ convergence protocol initialize --taker-fee=1 --maker-fee=0 --collateral-mint=$U
 convergence risk-engine initialize
 
 convergence protocol add-instrument --instrument-program=$SPOT_INSTRUMENT                --can-be-used-as-quote=true --validate-data-account-amount=1 --prepare-to-settle-account-amount=7 --settle-account-amount=3 --revert-preparation-account-amount=3 --clean-up-account-amount=4 
-convergence protocol add-instrument --instrument-program=$PSYOPTIONS_EUROPEAN_INSTRUMENT --can-be-used-as-quote=true --validate-data-account-amount=2 --prepare-to-settle-account-amount=7 --settle-account-amount=3 --revert-preparation-account-amount=3 --clean-up-account-amount=4
-convergence protocol add-instrument --instrument-program=$PSYOPTIONS_AMERICAN_INSTRUMENT --can-be-used-as-quote=true --validate-data-account-amount=3 --prepare-to-settle-account-amount=7 --settle-account-amount=3 --revert-preparation-account-amount=3 --clean-up-account-amount=4
+convergence protocol add-instrument --instrument-program=$PSYOPTIONS_EUROPEAN_INSTRUMENT --can-be-used-as-quote=false --validate-data-account-amount=2 --prepare-to-settle-account-amount=7 --settle-account-amount=3 --revert-preparation-account-amount=3 --clean-up-account-amount=4
+convergence protocol add-instrument --instrument-program=$PSYOPTIONS_AMERICAN_INSTRUMENT --can-be-used-as-quote=false --validate-data-account-amount=3 --prepare-to-settle-account-amount=7 --settle-account-amount=3 --revert-preparation-account-amount=3 --clean-up-account-amount=4
 
 convergence protocol add-print-trade-provider --print-trade-provider-program $HXRO_PRINT_TRADE_PROVIDER --settlement-can-expire false
 convergence hxro initialize-config --valid-mpg $HXRO_MPG
@@ -29,6 +29,6 @@ convergence risk-engine set-risk-categories-info --new-value="0.05,5.0,0.10,1.0,
 convergence protocol add-base-asset --ticker=BTC --oracle-address=$BTC_ORACLE_ADDRESS --oracle-source=switchboard 
 convergence protocol add-base-asset --ticker=SOL --oracle-address=$SOL_ORACLE_ADDRESS --oracle-source=switchboard
 
-convergence protocol register-mint --mint=$BTC_MINT_DEVNET --base-asset-index=0 --rpc-endpoint=$RPC_ENDPOINT
-convergence protocol register-mint --mint=$SOL_MINT_DEVNET --base-asset-index=1 --rpc-endpoint=$RPC_ENDPOINT
-convergence protocol register-mint --mint=$USDC_MINT_DEVNET --rpc-endpoint=$RPC_ENDPOINT
+convergence protocol register-mint --mint=$BTC_MINT --base-asset-index=0 
+convergence protocol register-mint --mint=$SOL_MINT --base-asset-index=1
+convergence protocol register-mint --mint=$USDC_MINT