+++ packages/js/src/plugins/rfqModule/RfqClient.ts
@@ -82,8 +82,17 @@ import {
   createPrintTradeRfqOperation,
   preparePrintTradeSettlementOperation,
   PreparePrintTradeSettlementInput,
+  getSettlementResultHandler,
+  GetResponseStateAndActionInput,
+  getResponseStateAndActionOperation,
+  getResponseStateAndActionHandler,
 } from './operations';
 import { Response } from './models/Response';
+import {
+  GetRfqStateAndActionInput,
+  getRfqStateAndActionHandler,
+  getRfqStateAndActionOperation,
+} from './operations/getRfqStateAndAction';
 
 /**
  * This is a client for the Rfq module.
@@ -505,12 +514,26 @@ export class RfqClient {
   }
 
   /** {@inheritDoc getSettlementResultOperation} */
-  getSettlementResult(
-    input: GetSettlementResultInput,
-    options?: OperationOptions
-  ) {
-    return this.convergence
-      .operations()
-      .execute(getSettlementResultOperation(input), options);
+  getSettlementResult(input: GetSettlementResultInput) {
+    return getSettlementResultHandler.handle(
+      getSettlementResultOperation(input),
+      this.convergence
+    );
+  }
+
+  /** {@inheritDoc getResponseStateAndAction} */
+  getResponseStateAndAction(input: GetResponseStateAndActionInput) {
+    return getResponseStateAndActionHandler.handle(
+      getResponseStateAndActionOperation(input),
+      this.convergence
+    );
+  }
+
+  /** {@inheritDoc getRFqStateAndAction} */
+  getRfqStateAndAction(input: GetRfqStateAndActionInput) {
+    return getRfqStateAndActionHandler.handle(
+      getRfqStateAndActionOperation(input),
+      this.convergence
+    );
   }
 }