+++ packages/js/src/plugins/rfqModule/operations/index.ts
@@ -37,3 +37,4 @@ export * from './unlockResponseCollateral';
 export * from './unlockResponsesCollateral';
 export * from './unlockRfqsCollateral';
 export * from './getSettlementResult';
+export * from './getResponseStateAndAction';