+++ packages/js/src/plugins/rfqModule/plugin.ts
@@ -77,8 +77,14 @@ import {
   createPrintTradeRfqOperationHandler,
   preparePrintTradeSettlementOperationHandler,
   preparePrintTradeSettlementOperation,
+  getResponseStateAndActionOperation,
+  getResponseStateAndActionHandler,
 } from './operations';
 import { rfqProgram } from './program';
+import {
+  getRfqStateAndActionHandler,
+  getRfqStateAndActionOperation,
+} from './operations/getRfqStateAndAction';
 
 /** @group Plugins */
 export const rfqModule = (): ConvergencePlugin => ({
@@ -194,6 +200,11 @@ export const rfqModule = (): ConvergencePlugin => ({
       unlockRfqsCollateralOperationHandler
     );
     op.register(getSettlementResultOperation, getSettlementResultHandler);
+    op.register(
+      getResponseStateAndActionOperation,
+      getResponseStateAndActionHandler
+    );
+    op.register(getRfqStateAndActionOperation, getRfqStateAndActionHandler);
 
     convergence.rfqs = function () {
       return new RfqClient(this);