+++ packages/js/tests/unit/rfq.spec.ts
@@ -55,11 +55,15 @@ describe('unit.rfq', () => {
     expect(rfqs.length).toBeGreaterThan(0);
   });
 
-  // TODO ADD getRfqState function
+  // TODO ADD getRfqStateAndAction function
   it('cancel', async () => {
+    // Error Number: 6016. Error Message: Rfq is not in required state.
     const iterator: any = takerCvg.rfqs().findRfqs({});
     const rfqs = (await getAll(iterator)).flat().filter((rfq: any) => {
-      return rfq.state === 'active' && rfq.totalResponses === 0;
+      return (
+        takerCvg.rfqs().getRfqStateAndAction({ rfq, caller: 'taker' })
+          .rfqAction === 'Cancel'
+      );
     });
     expect(rfqs.length).toBeGreaterThan(0);
     const { responses } = await takerCvg
@@ -71,7 +75,10 @@ describe('unit.rfq', () => {
   it('unlock', async () => {
     const iterator: any = takerCvg.rfqs().findRfqs({});
     const rfqsBefore = (await getAll(iterator)).flat().filter((rfq: any) => {
-      return rfq.state === 'canceled' && rfq.totalResponses === 0;
+      return (
+        takerCvg.rfqs().getRfqStateAndAction({ rfq, caller: 'taker' })
+          .rfqAction === 'UnlockCollateral'
+      );
     });
     expect(rfqsBefore.length).toBeGreaterThan(0);
     const { responses } = await takerCvg.rfqs().unlockRfqsCollateral({
@@ -79,7 +86,10 @@ describe('unit.rfq', () => {
     });
     expect(responses.length).toBe(rfqsBefore.length);
     const rfqsAfter = (await getAll(iterator)).flat().filter((rfq: any) => {
-      return rfq.state === 'canceled' && rfq.totalResponses === 0;
+      return (
+        takerCvg.rfqs().getRfqStateAndAction({ rfq, caller: 'taker' })
+          .rfqAction === 'Cleanup'
+      );
     });
     rfqsAfter.map((rfq: any) => {
       expect(rfq.totalTakerCollateralLocked).toBe(0);
@@ -89,7 +99,10 @@ describe('unit.rfq', () => {
   it('clean up', async () => {
     const iterator: any = takerCvg.rfqs().findRfqs({});
     const rfqs = (await getAll(iterator)).flat().filter((rfq: any) => {
-      return rfq.state === 'canceled';
+      return (
+        takerCvg.rfqs().getRfqStateAndAction({ rfq, caller: 'taker' })
+          .rfqAction === 'Cleanup'
+      );
     });
     expect(rfqs.length).toBeGreaterThan(0);
     await takerCvg.rfqs().cleanUpRfqs({