+++ packages/js/tests/unit/riskEngine.spec.ts
@@ -1,6 +1,4 @@
 import { expect } from 'expect';
-// TODO: Should not be using this
-import { QuoteSide } from '@convergence-rfq/rfq';
 
 import { DEFAULT_RISK_CATEGORIES_INFO } from '../../src';
 import {
@@ -157,7 +155,7 @@ describe('unit.riskEngine', () => {
       .calculateCollateralForConfirmation({
         rfqAddress: rfq.address,
         responseAddress: rfqResponse.address,
-        confirmation: { side: QuoteSide.Ask, overrideLegMultiplier: 1 },
+        confirmation: { side: 'bid', overrideLegMultiplier: 1 },
       });
     expect(requiredCollateral).toBeCloseTo(rfqResponse.makerCollateralLocked);
   });