+++ packages/js/tests/unit/settlementResult.spec.ts
@@ -1,6 +1,5 @@
 import expect from 'expect';
 import { Mint } from '@solana/spl-token';
-import { QuoteSide } from '@convergence-rfq/rfq';
 import {
   createAmericanCoveredCallRfq,
   createEuropeanCoveredCallRfq,
@@ -44,11 +43,11 @@ describe('unit.settlementResult', () => {
     );
     expect(rfqResponse).toHaveProperty('address');
 
-    const responseResult = await takerCvg.rfqs().getSettlementResult({
+    const responseResult = takerCvg.rfqs().getSettlementResult({
       rfq,
       response: rfqResponse,
       confirmation: {
-        side: rfqResponse?.ask ? QuoteSide.Ask : QuoteSide.Bid,
+        side: rfqResponse?.ask ? 'ask' : 'bid',
       },
     });
 
@@ -68,7 +67,7 @@ describe('unit.settlementResult', () => {
     const refreshedResponse = await takerCvg.rfqs().findResponseByAddress({
       address: rfqResponse.address,
     });
-    const confimedResponseResult = await takerCvg.rfqs().getSettlementResult({
+    const confimedResponseResult = takerCvg.rfqs().getSettlementResult({
       rfq,
       response: refreshedResponse,
     });
@@ -100,11 +99,11 @@ describe('unit.settlementResult', () => {
       undefined,
       pricePerToken
     );
-    const responseResult = await takerCvg.rfqs().getSettlementResult({
+    const responseResult = takerCvg.rfqs().getSettlementResult({
       rfq,
       response: rfqResponse,
       confirmation: {
-        side: rfqResponse?.ask ? QuoteSide.Ask : QuoteSide.Bid,
+        side: rfqResponse?.ask ? 'ask' : 'bid',
       },
     });
 
@@ -122,7 +121,7 @@ describe('unit.settlementResult', () => {
     const refreshedResponse = await takerCvg.rfqs().findResponseByAddress({
       address: rfqResponse.address,
     });
-    const confimedResponseResult = await takerCvg.rfqs().getSettlementResult({
+    const confimedResponseResult = takerCvg.rfqs().getSettlementResult({
       rfq,
       response: refreshedResponse,
     });
@@ -147,11 +146,11 @@ describe('unit.settlementResult', () => {
       7.456
     );
     expect(rfqResponse).toHaveProperty('address');
-    const responseResult = await takerCvg.rfqs().getSettlementResult({
+    const responseResult = takerCvg.rfqs().getSettlementResult({
       rfq,
       response: rfqResponse,
       confirmation: {
-        side: rfqResponse?.ask ? QuoteSide.Ask : QuoteSide.Bid,
+        side: rfqResponse?.ask ? 'ask' : 'bid',
       },
     });
 
@@ -168,7 +167,7 @@ describe('unit.settlementResult', () => {
     const refreshedResponse = await takerCvg.rfqs().findResponseByAddress({
       address: rfqResponse.address,
     });
-    const confimedResponseResult = await takerCvg.rfqs().getSettlementResult({
+    const confimedResponseResult = takerCvg.rfqs().getSettlementResult({
       rfq,
       response: refreshedResponse,
     });
@@ -192,11 +191,11 @@ describe('unit.settlementResult', () => {
       undefined,
       8.456123456
     );
-    const responseResult = await takerCvg.rfqs().getSettlementResult({
+    const responseResult = takerCvg.rfqs().getSettlementResult({
       rfq,
       response: rfqResponse,
       confirmation: {
-        side: rfqResponse?.ask ? QuoteSide.Ask : QuoteSide.Bid,
+        side: rfqResponse?.ask ? 'ask' : 'bid',
       },
     });
 
@@ -215,7 +214,7 @@ describe('unit.settlementResult', () => {
     const refreshedResponse = await takerCvg.rfqs().findResponseByAddress({
       address: rfqResponse.address,
     });
-    const confimedResponseResult = await takerCvg.rfqs().getSettlementResult({
+    const confimedResponseResult = takerCvg.rfqs().getSettlementResult({
       rfq,
       response: refreshedResponse,
     });
@@ -230,7 +229,7 @@ describe('unit.settlementResult', () => {
     const baseAmount = 5.632123779;
     const quoteAmount1 = 24_300.75;
     const quoteAmount2 = 28_899.75;
-    const responseSide = QuoteSide.Bid;
+    const responseSide = 'bid';
     const { rfq } = await createRfq(takerCvg, baseAmount, 'two-way');
     expect(rfq).toHaveProperty('address');
 
@@ -241,7 +240,7 @@ describe('unit.settlementResult', () => {
       quoteAmount2
     );
     expect(rfqResponse).toHaveProperty('address');
-    const responseResult = await takerCvg.rfqs().getSettlementResult({
+    const responseResult = takerCvg.rfqs().getSettlementResult({
       rfq,
       response: rfqResponse,
       confirmation: {
@@ -262,7 +261,7 @@ describe('unit.settlementResult', () => {
     const refreshedResponse = await takerCvg.rfqs().findResponseByAddress({
       address: rfqResponse.address,
     });
-    const confimedResponseResult = await takerCvg.rfqs().getSettlementResult({
+    const confimedResponseResult = takerCvg.rfqs().getSettlementResult({
       rfq,
       response: refreshedResponse,
     });
@@ -276,7 +275,7 @@ describe('unit.settlementResult', () => {
     const baseAmount = 9.632123779;
     const quoteAmount1 = 14_300.75;
     const quoteAmount2 = 18_899.75;
-    const responseSide = QuoteSide.Bid;
+    const responseSide = 'bid';
     const { rfq } = await createRfq(takerCvg, baseAmount, 'two-way');
     expect(rfq).toHaveProperty('address');
 
@@ -287,7 +286,7 @@ describe('unit.settlementResult', () => {
       quoteAmount2
     );
     expect(rfqResponse).toHaveProperty('address');
-    const responseResult = await takerCvg.rfqs().getSettlementResult({
+    const responseResult = takerCvg.rfqs().getSettlementResult({
       rfq,
       response: rfqResponse,
       confirmation: {
@@ -308,7 +307,7 @@ describe('unit.settlementResult', () => {
     const refreshedResponse = await takerCvg.rfqs().findResponseByAddress({
       address: rfqResponse.address,
     });
-    const confimedResponseResult = await takerCvg.rfqs().getSettlementResult({
+    const confimedResponseResult = takerCvg.rfqs().getSettlementResult({
       rfq,
       response: refreshedResponse,
     });
@@ -329,11 +328,11 @@ describe('unit.settlementResult', () => {
 
     const { rfqResponse } = await respondToRfq(makerCvg, rfq, 12.1);
     expect(rfqResponse).toHaveProperty('address');
-    const responseResult = await takerCvg.rfqs().getSettlementResult({
+    const responseResult = takerCvg.rfqs().getSettlementResult({
       rfq,
       response: rfqResponse,
       confirmation: {
-        side: rfqResponse?.ask ? QuoteSide.Ask : QuoteSide.Bid,
+        side: rfqResponse?.ask ? 'ask' : 'bid',
       },
     });
 
@@ -355,7 +354,7 @@ describe('unit.settlementResult', () => {
     const refreshedResponse = await takerCvg.rfqs().findResponseByAddress({
       address: rfqResponse.address,
     });
-    const confimedResponseResult = await takerCvg.rfqs().getSettlementResult({
+    const confimedResponseResult = takerCvg.rfqs().getSettlementResult({
       rfq,
       response: refreshedResponse,
     });
@@ -378,11 +377,11 @@ describe('unit.settlementResult', () => {
     expect(response.signature).toBeDefined();
 
     const { rfqResponse } = await respondToRfq(makerCvg, rfq, 12.34);
-    const responseResult = await takerCvg.rfqs().getSettlementResult({
+    const responseResult = takerCvg.rfqs().getSettlementResult({
       rfq,
       response: rfqResponse,
       confirmation: {
-        side: rfqResponse?.ask ? QuoteSide.Ask : QuoteSide.Bid,
+        side: rfqResponse?.ask ? 'ask' : 'bid',
       },
     });
 
@@ -402,7 +401,7 @@ describe('unit.settlementResult', () => {
     const refreshedResponse = await takerCvg.rfqs().findResponseByAddress({
       address: rfqResponse.address,
     });
-    const confimedResponseResult = await takerCvg.rfqs().getSettlementResult({
+    const confimedResponseResult = takerCvg.rfqs().getSettlementResult({
       rfq,
       response: refreshedResponse,
     });